refactor: relocate client, server, and model packages under internal
