+++ internal/client/listener.go
@@ -1,4 +1,4 @@
-package internal
+package client
 
 import (
 	"fmt"
