+++ internal/model/resultData.go
