+++ internal/model/sensorData.go
